hw lesson_6 level_2

+++ src/main/java/org/javaguru/student_alexey_yakovlev/lesson_6/homework/level_1_intern/task_2/AverageFinder.java
@@ -2,9 +2,9 @@ package org.javaguru.student_alexey_yakovlev.lesson_6.homework.level_1_intern.ta
 
 class AverageFinder {
 
-    public int findAverage(int num1, int num2) {
+    public double findAverage(int num1, int num2) {
         int sum = num1 + num2;
-        return sum / 2;
+        return sum / (double)2;
     }
 
 }
